groups

+++ server/src/users/services/groups/groups.service.ts
@@ -1,5 +1,6 @@
 import {
   BadRequestException,
+  ForbiddenException,
   Inject,
   Injectable,
   InternalServerErrorException,
@@ -9,11 +10,14 @@ import { Model, Types } from "mongoose";
 import { Group } from "src/schemas/group.schema";
 import { CreateGroupDto } from "src/users/dtos/CreateGroup.dto";
 import { UsersService } from "../users/users.service";
+import { User } from "src/schemas/user.schema";
+import { PatchGroupDto } from "src/users/dtos/PatchGroup.dto";
 
 @Injectable()
 export class GroupsService {
   constructor(
     @InjectModel(Group.name) private groupModel: Model<Group>,
+    @InjectModel(User.name) private userModel: Model<User>,
     @Inject("USER_SERVICE") private readonly userService: UsersService
   ) {}
 
@@ -46,9 +50,17 @@ export class GroupsService {
   }
 
   async patchGroup(
-    groupDto: CreateGroupDto,
-    groupId: Types.ObjectId
+    groupDto: PatchGroupDto,
+    groupId: Types.ObjectId,
+    userId: Types.ObjectId
   ): Promise<Group> {
+    const group = await this.groupModel.findById(groupId);
+
+    if (!group) throw new BadRequestException("Nie znaleziono podanej grupy.");
+
+    if (!this.userService.isUserAllowed(userId, group.owner))
+      throw new ForbiddenException();
+
     if (groupDto.owner) {
       const owner = await this.userService.findUserByUsername(
         groupDto.owner.toString()
@@ -61,22 +73,101 @@ export class GroupsService {
 
       groupDto.owner = owner._id;
     }
+
     try {
-      return await this.groupModel.findByIdAndUpdate(groupId, groupDto, {
-        returnOriginal: false,
-      });
+      return await group.updateOne(groupDto, { returnOriginal: false });
     } catch {
       throw new InternalServerErrorException("Edycja grupy nie powiodła się.");
     }
   }
 
-  async deleteGroup(groupId: Types.ObjectId): Promise<void> {
+  async deleteGroup(
+    groupId: Types.ObjectId,
+    userId: Types.ObjectId
+  ): Promise<void> {
     const group = await this.groupModel.findById(groupId);
 
     if (!group) throw new BadRequestException("Nie znaleziono podanej grupy.");
 
+    if (!this.userService.isUserAllowed(userId, group.owner))
+      throw new ForbiddenException();
+
     try {
       await group.deleteOne();
+    } catch {
+      throw new InternalServerErrorException(
+        "Usuwanie grupy nie powiodło się."
+      );
+    }
+  }
+
+  async addUserToGroup(
+    groupId: Types.ObjectId,
+    userToAdd: string,
+    userId: Types.ObjectId
+  ) {
+    const group = await this.groupModel.findById(groupId);
+    if (!group) throw new BadRequestException("Nie znaleziono podanej grupy.");
+
+    if (!this.userService.isUserAllowed(userId, group.owner))
+      throw new ForbiddenException();
+
+    const user = await this.userModel.findOne({ username: userToAdd });
+    if (!user)
+      throw new BadRequestException("Użytkownik o takim id nie istnieje.");
+
+    if (group.owner.equals(user._id))
+      throw new BadRequestException(
+        "Nie możesz dodać właściciela do jego własnej grupy."
+      );
+
+    if (group.members.includes(user._id))
+      throw new BadRequestException(
+        "Użytkownik o podanej nazwie jest już w grupie."
+      );
+
+    group.members.push(user._id);
+    user.groups.push(group._id);
+
+    try {
+      await group.save();
+      await user.save();
+    } catch {
+      throw new InternalServerErrorException(
+        "Usuwanie grupy nie powiodło się."
+      );
+    }
+  }
+
+  async removeUserFromGroup(
+    groupId: Types.ObjectId,
+    userToRemove: string,
+    userId: Types.ObjectId
+  ) {
+    const group = await this.groupModel.findById(groupId);
+    if (!group) throw new BadRequestException("Nie znaleziono podanej grupy.");
+
+    if (!this.userService.isUserAllowed(userId, group.owner))
+      throw new ForbiddenException();
+
+    const user = await this.userModel.findOne({ username: userToRemove });
+    if (!user)
+      throw new BadRequestException("Użytkownik o takim id nie istnieje.");
+
+    if (group.owner.equals(user._id))
+      throw new BadRequestException(
+        "Nie możesz usunąć właściciela z jego własnej grupy."
+      );
+
+    if (!group.members.includes(user._id))
+      throw new BadRequestException("Takiego użytkownika nie ma w grupie.");
+
+    group.members = group.members.filter((_user) => !_user.equals(user._id));
+    user.groups = user.groups.filter((_group) => !_group.equals(group._id));
+
+    try {
+      await group.save();
+      await user.save();
     } catch {
       throw new InternalServerErrorException(
         "Usuwanie grupy nie powiodło się."
